feat: versão aprimorada
