+++ VersãoAnterior/gpt_code_original.py
