+++ VersãoAnterior/gpt_code_updated.py
