+++ VersãoAprimorada/readme.md
@@ -1,0 +1,7 @@
+ # ARQUIVO PARA LEITURA DO PROJETO
+
+```
+Entre na pasta "VersãoAprimorada" usando "cd VersãoAprimorada", no terminal.
+
+Para rodar o projeto, basta digitar "python3 index.py" ou qualquer outro comando que rode python.
+```
